Add files via upload
java program that meets the following parameters. 

Create a linked list called Rainbow.
Load the colors of the Rainbow into the linked list (Remember: ROYGBIV)
Display the number of items in the list and print its values to the screen
Add an item to the head of the list by asking the user for an item to add at the beginning. Display the updated list
Add an item to the end of the list by asking the user for an item to add at the end. Display the updated list
Prompt the user for an item to remove from the list, make sure that it exists before attempting to remove it, and also ensure you check to make sure the list is not empty. Display the updated list
Ask the user to either exit or repeat removing items. If the user chooses to exit, quit the program. If they choose to repeat removing elements return to step 6.